Failing test for case where isTextArea property of TextFieldView changes.
The issue is that when the property changes, the code necessarily updates the HTML but it attempts to attach events based on the HTML present in the DOM, resulting in a failure to wire up the change event properly. This case never arises if isTextArea is set statically at the time the TextFieldView is created. It arises if the property is changed after creation of the view, either manually or via a binding.