Fix for the issue described in f382e38b3b94 related to how TextFieldView handles a change to isTextArea. See extended commit message for potential breaking change.
As part of the bug fix, this change refactors responsibility for isTextArea completely into TextFieldView. Previously, it was awkwardly split by SC.FieldView and SC.TextFieldView. The change is fully compatible with the internal API, but if you have any application views that extend SC.FieldView and depend on the incomplete isTextArea logic in SC.FieldView, you will need to either extend from TextFieldView instead or implement your own support for isTextArea.